feat(landingpage-new): implement code hike and change directory
