+++ packages/landingpage-new/mdx/react-guide.mdx
@@ -1,4 +1,18 @@
 # Create your first Project
+<CH.Section>
+
+## Code Links Example
+
+We are looking at the [console.log](focus://console.js#2) function today
+
+<CH.Code showCopyButton={true}>
+```js console.js
+console.table(["Hello", "World"])
+console.log("Hello World")
+```
+</CH.Code>
+
+</CH.Section>
 
 **Let's get you up to speed quickly with the usage of the Elements to create a simple login page.**
