+++ packages/landingpage-new/package.json
@@ -11,6 +11,7 @@
     "clean": "shx rm -rf out && shx rm -rf .next"
   },
   "dependencies": {
+    "@code-hike/mdx": "^0.7.4",
     "@inovex.de/elements-react": "^7.1.2",
     "@mdx-js/loader": "^2.1.4",
     "@mdx-js/react": "^2.1.4",
